refactor(admin): migrate gateway consumers to shared types
