+++ src/components/providers/signal/SetupPanel.vue
@@ -81,8 +81,8 @@ import {
 	cancelInteractiveSetup,
 	interactiveSetupStep,
 	startInteractiveSetup,
-	type InteractiveSetupResponse,
 } from '../../../services/adminGatewayApi.ts'
+import type { InteractiveSetupResponse } from '../../../services/adminGatewayTypes.ts'
 
 export default defineComponent({
 	name: 'SetupPanel',
