+++ src/components/providers/whatsappbusiness/SetupPanel.vue
@@ -192,8 +192,8 @@ import {
 	cancelInteractiveSetup,
 	interactiveSetupStep,
 	startInteractiveSetup,
-	type InteractiveSetupResponse,
 } from '../../../services/adminGatewayApi.ts'
+import type { InteractiveSetupResponse } from '../../../services/adminGatewayTypes.ts'
 
 type PhoneNumberOption = {
 	id: string
