+++ src/tests/components/GatewayInstanceCard.spec.ts
@@ -5,7 +5,7 @@ import { describe, expect, it, vi } from 'vitest'
 import { mount } from '@vue/test-utils'
 import { defineComponent } from 'vue'
 import GatewayInstanceCard from '../../components/GatewayInstanceCard.vue'
-import type { FieldDefinition, GatewayInstance } from '../../services/adminGatewayApi.ts'
+import type { FieldDefinition, GatewayInstance } from '../../services/adminGatewayTypes.ts'
 
 Object.defineProperty(window, 'matchMedia', {
 	writable: true,
