+++ src/tests/components/GatewayInstanceModal.spec.ts
@@ -5,7 +5,7 @@ import { describe, expect, it, vi } from 'vitest'
 import { flushPromises, mount } from '@vue/test-utils'
 import { defineComponent } from 'vue'
 import GatewayInstanceModal from '../../components/GatewayInstanceModal.vue'
-import type { GatewayInfo } from '../../services/adminGatewayApi.ts'
+import type { GatewayInfo } from '../../services/adminGatewayTypes.ts'
 
 Object.defineProperty(window, 'matchMedia', {
 	writable: true,
